% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asMuiTree.R
\name{asMuiTree}
\alias{asMuiTree}
\title{Create choice structure for [RichTreeView()]}
\usage{
asMuiTree(data, levels = names(data), levels_id = NULL, ...)
}
\arguments{
\item{data}{A `data.frame`.}

\item{levels}{Variables identifying hierarchical levels,
values of those variables will be used as text displayed.}

\item{levels_id}{Variable to use as ID for nodes. Careful! Spaces are not allowed in IDs.}

\item{...}{Addtional arguments.}
}
\value{
a `list` that can be used in [RichTreeView()].
}
\description{
This function is adapted from `shinyWidgets::create_tree()`. A big thank you
to the authors of the shinyWidgets for creating the `create_tree()` function.
}
\seealso{
`shinyWidgets::create_tree()`
}
