% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/muiTreeDependency.R
\name{muiTreeViewDependency}
\alias{muiTreeViewDependency}
\title{Mui X Tree View JS dependency}
\usage{
muiTreeViewDependency()
}
\value{
HTML dependency object.
}
\description{
Mui X Tree View JS dependency
}
