% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R, R/documentation.R
\name{TreeItem}
\alias{TreeItem}
\title{TreeItem}
\usage{
TreeItem(...)
}
\arguments{
\item{...}{Props to pass to the component.}
}
\value{
Object with \code{shiny.tag} class suitable for use in the UI of a Shiny app.
}
\description{
\url{https://mui.com/x/api/tree-view/tree-item/}
}
\details{
\itemize{
\item itemId \code{string} \cr Default is - The id of the item. Must be unique.
\item children \code{any} \cr Default is - The content of the component.
\item classes \code{object} \cr Default is - Override or extend the styles applied to the component.See CSS classes API below for more details.
\item disabled \code{bool} \cr Default is FALSE If true, the item is disabled.
\item id \code{string} \cr Default is - The id attribute of the item. If not provided, it will be generated.
\item label \code{node} \cr Default is - The label of the item.
\item onBlur \code{func} \cr Default is - Callback fired when the item root is blurred.
\item onFocus \code{unsupportedProp} \cr Default is - This prop isn't supported. Use the onItemFocus callback on the tree if you need to monitor an item's focus.
\item onKeyDown \code{func} \cr Default is - Callback fired when a key is pressed on the keyboard and the tree is in focus.
\item slotProps \code{object} \cr Default is  The props used for each component slot.
\item slots \code{object} \cr Default is  Overridable component slots.See Slots API below for more details.
\item sx \verb{Array<func| object| bool>| func| object} \cr Default is - The system prop that allows defining system overrides as well as additional CSS styles.See the \code{sx} page for more details.
}
}
